% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biEmcdf.R
\name{Biemcdf}
\alias{Biemcdf}
\title{Computes bivariate empirical joint distribution}
\usage{
Biemcdf(data)
}
\arguments{
\item{data}{a numeric matrix stores data.}
}
\value{
a numeric as value of empirical joint CDF function.
}
\description{
This function computes empirical joint distribution (joint CDF) table with dynamical programming.
}
\details{
When data is a numeric matrix, this function computes joint empirical CDF with single thread.
When data is an object of class "emcdf_obj", it computes with multi-thread. Parameter "a"
must have equal length as the column number of data. Both single-thread and multi-thread emcdf
algorithms are faster than using the bulit-in function sum\{base\}. See example for simulation.
Note that initializing threads and spliting data takes time though it's a one-time task.
Thus for big data, big number of CDF computation, multi-thread is recommended. Yet for small data,
small number of CDF computation, single thread is faster.
}
\examples{
n = 10^6
set.seed(123)
x = rnorm(n)
y = rnorm(n)
data = cbind(x, y)

}
\author{
En-shuo Hsu
}

