% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNum.R
\name{coreNum}
\alias{coreNum}
\title{CPU core number}
\usage{
coreNum()
}
\value{
an integer indicating the total number of cores in CPU.
}
\description{
This function reports the total core number in CPU on the machine.
}
\details{
This function is a wrapper for the C++ function std::thread::hardware_concurrency();
If core number is not detected, 0 will be returned. See "See Also" for details.
}
\examples{

num = coreNum()

}
\author{
En-shuo Hsu
}
\seealso{
http://www.cplusplus.com/reference/thread/thread/hardware_concurrency/
}

